Information for uploaded image is now being place in database. #19.
Create join tables to join images table with cars, people and licences. This way the images table
doesn't care what the image is; I have one central images table and everyting else joins to it.

I need fix a couple of issue on the front end and the upload will be complete.